updated the jwt-editor files
